✨ Feat: 가게의 리뷰 목록 조회 api 구현
week9 - 실습

+++ src/main/java/umc/spring/domain/Review.java
@@ -28,7 +28,7 @@ public class Review extends BaseEntity {
     private String description;
 
     @Column(nullable = false)
-    private Integer stars;
+    private Float stars;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
